requirements.txt
Django>~2.0
djangorestframework>~3.7.3
factory-boy>~2.9.2
coreapi-cli>~1.0
django-cors-headers
Pillow
django-markdownx
dry_rest_permissions>~0.1.10
django-guardian>~1.4.9
django-admin-sortable2>=0.6.19
pymdown-extensions>=4.8
psycopg2
dj-database-url
whitenoise
django-heroku
